requirements.txt
streamlit>=1.37.0
litellm>=1.40.0
python-dateutil>=2.8.2
pandas>=2.0.0
psycopg2-binary>=2.9.9  # For Supabase PostgreSQL connection (Streamlit's sql connection uses this under the hood)


# Optional but recommended for better performance / caching
cachetools>=5.3.0
dotenv
sqlalchemy
